Create custom USB VCP library and RTOS Logging Framework (#68)
Logger.h defines the logging functions that serializes access to the USB interface. 

ts_printf2() defines a non-threaded approach to serializing messages to the USB interface by acquiring a mutex whenever a thread attempts to send a message over USB. This is slower and blocking, causing threads to wait for the USB interface, prefer using ts_printf() and the logging framework where possible.

* Create custom USB VCP library

* Create tests for USB base library

Access to this buffer will be managed in an RTOS environment with a Semaphore.

* Working thread-safe USB library