CAN Autogeneration Library
Automatically takes in a CSV of our own format, parses it, and then generates C code for packing/unpacking the structs. This struct will be used by longhorn lib's CAN interface. Note that currently this library is using the Nightwatch's CAN packet data for testing. This will be replaced with the 2026 car's own packets later on as that gets decided.