Create RTOS CAN Library (#113)
* feat: Add thread-safe CAN RTOS wrapper

Adds a new RTOS wrapper for the Longhorn CAN library.
- Implements `drivers/longhorn-lib/rtos/can.c` and `can.h`.
- Provides thread-safe versions of registration and sending functions using FreeRTOS mutexes.
- Implements a high-priority transceiver task for periodic message sending.
- Implements a receiver task that offloads message unpacking from the ISR to a thread using a queue.
- Ensures no race conditions or deadlocks by correctly separating ISR and thread contexts.

* Working CAN lib for RTOS

Also automatically sets the send task for the library to be built based on the GCD of the CAN packets. This ensures that messages can be sent on strict deadlines while the CAN system does not need to utilize more CPU than it already does.

* Allow CAN Bus staggering, reduces load spikes

* Update send test to account for phasing

* Fix build, C++ 👀

* Don't allow 0 size DLCs, other cleanup

* Clarify malloc and free usage in README